Create README.md
An application that fetches data from an API resource and displays the data in the API., the application has a homepage that explains a  bit of how the application helps the user to view the ships and some image of the ship. After the home page the user will be directed to a recycle view where he will have to view all the ships row which consist of only an image and name of ship and  also the homeport of the ship while scrolling and he can learn much  about a ship by pressing on one of the recycler view to get to know more details about the ship.